cli: Fix ambiguous references
This patch fixes ambiguous references. Java 8 introduced Target-type
Inference which supports to infer a type not only from assignment,
but also from method calls. For example:

 put(get(value);

Java 8 is now able to infer type, and call put with proper type,
in Java 7 the Object type was always used.

Since Formatter class is using ambigous references, we must specify a
return types of the ExtMap calls, so we avoid ambiguous reference
exceptions.

Bug-Url: https://bugzilla.redhat.com/1415704
Change-Id: I65419109b1e5ee6e8dc9068c24edad235ddff5d3
Signed-off-by: Ondra Machacek <omachace@redhat.com>